support added for vmware
The same image that I'm building for virtualbox is now able to be built
on vmware fusion.  There is a timeout error where packer doesn't know
that vmware has shutdown the image on a halt, however that could be
because i'm running a new version of fusion v8 that just came out.